add cities in mainscreen
